requirements.txt
# ── Core ────────────────────────────────────────
numpy>=1.24,<2.0
pandas>=2.0
scipy>=1.11
scikit-learn>=1.3
matplotlib>=3.7
tqdm>=4.65
pydantic>=2.0
python-dotenv>=1.0

# ── Deep Learning ───────────────────────────────
torch>=2.1
torch-geometric>=2.4

# ── RAG / NLP ───────────────────────────────────
langchain>=0.2
langchain-community>=0.2
langchain-text-splitters>=0.2
chromadb>=0.4
sentence-transformers>=2.2
pymupdf>=1.23            # PDF parsing (fitz)

# ── LLM Providers ──────────────────────────────
groq>=0.4
google-generativeai>=0.5
openai>=1.12
anthropic>=0.25
httpx>=0.27

# ── Dashboard ───────────────────────────────────
streamlit>=1.32
plotly>=5.18

# ── Testing ─────────────────────────────────────
pytest>=8.0
